Add rmdup_for_ase.py
Script from Ryo Kita to remove duplicate reads from a bam file. Randomizes quality scores so that the saved read is random, not necessarily the highest quality read. This prevents a bias favoring the read containing the reference allele, which will likely have a higher mapping quality than reads with the non-reference allele.